Move helpers out of lib.rs
